master

+++ ft_strlen.c
@@ -6,11 +6,12 @@
 /*   By: vnilprap <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/21 23:28:28 by vnilprap          #+#    #+#             */
-/*   Updated: 2022/02/23 16:21:33 by vnilprap         ###   ########.fr       */
+/*   Updated: 2022/03/02 21:21:41 by vnilprap         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-int	ft_strlen(char *c)
+size_t	ft_strlen(char *c)
 {
 	int	i;
